tree class created

+++ visual data structures/src/array_sorting.cpp
@@ -63,11 +63,11 @@ void array_sorting<T>::heaps(T*& arr, szt n, bool (*f)(type, type))
 {
 	// build
 	if (f == nullptr) f = [](type x, type y)->bool {return x > y; };
-	for (size_t i = n / 2 - 1; i >= 0; i--)
+	for (int i = n / 2 - 1; i >= 0; i--)
 		heapify(arr, n, i, f);
 
 	//sort
-	for (size_t i = n - 1; i >= 0; i--)
+	for (int i = n - 1; i >= 0; i--)
 	{
 		swap(arr, 0, i);
 		heapify(arr, i, 0, f);
